refactor: Move files to have a common directory architecture
The directory architecture correspond to the one used for the bootcamp.
