+++ app/src/main/java/com/github/se/project/ui/profile/CreateProfile.kt
@@ -1,4 +1,4 @@
-package com.github.se.project.ui.authentification
+package com.github.se.project.ui.profile
 
 import android.widget.Toast
 import androidx.compose.foundation.layout.Arrangement
